+++ docs/changelog/changelog.md
@@ -12,6 +12,7 @@
 - [Enable single zaak migration from Zaaktype page](https://dimpact.atlassian.net/browse/DATA-265)
 - [Configuration of Roltype per Zaaktype](https://dimpact.atlassian.net/browse/DATA-331)
 - [Extend PDF content: un-mapped properties](https://dimpact.atlassian.net/browse/DATA-386)
+- [Migrate Behandelaar (Rollen-related)](https://dimpact.atlassian.net/browse/DATA-332)
 
 ## v0.1.0